moved Upvote, Comment, Community widgets to DHiveUX
