+++ lib/hive_flutter_kit.dart
@@ -21,7 +21,7 @@ export 'ux/login_screen.dart';
 export 'ux/switch_user.dart';
 export 'ux/upvote_bottomsheet.dart';
 export 'ux/upvote.dart';
-export 'ux/community_list.dart';
+export 'ux/dhive/community_list/community_list.dart';
 export 'ux/edit_profile.dart';
 export 'ux/image_picker.dart';
 export 'ux/user_profile_bottom_sheet.dart';
